Change sectionNumber field to section_number
